chore: Bulk export of multiple documents
Widget as an extension of standard Polarion's table
