+++ src/main/resources/META-INF/hivemodule.xml
@@ -92,8 +92,5 @@
     <contribution configuration-id="com.polarion.alm.richpages.widgets">
         <widget id="ch.sbb.polarion.extension.pdf.exporter.widgets.bulkPdfExportWidget" instance="instance:ch.sbb.polarion.extension.pdf_exporter.widgets.BulkPdfExportWidget" />
     </contribution>
-    <contribution configuration-id="com.polarion.alm.richpages.widgets">
-        <widget id="ch.sbb.polarion.extension.pdf.exporter.widgets.bulkPdfExportTable" instance="instance:ch.sbb.polarion.extension.pdf_exporter.widgets.BulkPdfExportTable" />
-    </contribution>
 
 </module>
